Author and User controllers - create, update, delete

+++ module/Radio/src/Radio/Controller/Author.php
@@ -60,30 +60,42 @@ class Author extends BaseController {
     }
 
     public function create($data) {
-        // validation        
-        if ( !isset($data['name']) || !isset($data['photo']) ||
-        !isset($data['avatar']) || !isset($data['introduction']) ||
-        !isset($data['user']) ) {
-            $this->getResponse()->setStatusCode(400);
-            return new JsonModel(array("error" => "Mandatory fields: name, photo, avatar, introduction, user."));
-        }
-        
-        // validate user id
-        // TODO: validate via DB
-        if ( !is_numeric($data['user']) ) {
-            $this->getResponse()->setStatusCode(400);
-            return new JsonModel(array("error" => "user must be numeric."));
-        }
-        
         try {
+            // validation
+            if ( !isset($data['name']) || !isset($data['photo']) ||
+                            !isset($data['avatar']) || !isset($data['introduction']) ||
+                            !isset($data['user_id']) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "Mandatory fields: name, photo, avatar, introduction, user_id."));
+            }
+            
+            // validate user id
+            if ( !is_numeric($data['user_id']) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "User id must be numeric."));
+            }            
+            $user = $this->getEntityManager()->find('Radio\Entity\User', $data['user_id']);
+            if ( is_null($user) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "User id does not exist in DB."));
+            }
+            // check that user is not yet assigned to author
+            $check_author = $this->getEntityManager()
+                    ->getRepository('Radio\Entity\Author')
+                    ->findOneBy(array('user' => $data['user_id']));
+
+            if ( !is_null($check_author) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "User is already assigned to another author."));
+            }
+            
             $author = new \Radio\Entity\Author();
             
             $author->setName($data['name']);
             $author->setPhoto($data['photo']);
             $author->setAvatar($data['avatar']);
             $author->setIntroduction($data['introduction']);
-            $user = $this->getEntityManager()->find('Radio\Entity\User', $data['user']);
-    	    $author->setUser($user);
+            $author->setUser($user);
 
     	    $this->getEntityManager()->persist($author);
     	    $this->getEntityManager()->flush();
@@ -96,22 +108,36 @@ class Author extends BaseController {
     }
 
     public function update($id, $data) {
-        // validation        
-        if ( !isset($data['name']) && !isset($data['photo']) &&
-        !isset($data['avatar']) && !isset($data['introduction']) &&
-        !isset($data['user']) ) {
-            $this->getResponse()->setStatusCode(400);
-            return new JsonModel(array("error" => "One of the following fields must exist: name, photo, avatar, introduction, user."));
-        }
-        
-        // validate user id
-        // TODO: validate via DB
-        if ( isset($data['user']) && !is_numeric($data['user']) ) {
-            $this->getResponse()->setStatusCode(400);
-            return new JsonModel(array("error" => "user must be numeric."));
-        }
         try {
             $author = $this->getEntityManager()->find('Radio\Entity\Author', $id);
+
+            // validation
+            if ( is_null($author) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "Author id does not exist."));
+            }            
+            
+            if ( !isset($data['name']) && !isset($data['photo']) &&
+                            !isset($data['avatar']) && !isset($data['introduction']) &&
+                            !isset($data['user_id']) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "One of the following fields must exist: name, photo, avatar, introduction, user_id."));
+            }
+            
+            // validate user id
+            if ( isset($data['user_id']) && !is_numeric($data['user_id']) ) {
+                $this->getResponse()->setStatusCode(400);
+                return new JsonModel(array("error" => "User id must be numeric."));
+            }
+            $user = null;
+            if ( isset($data['user_id']) ) {
+                $user = $this->getEntityManager()->find('Radio\Entity\User', $data['user_id']);
+                if ( is_null($user) ) {
+                    return new JsonModel(array("error" => "User id does not exist in DB."));
+                }
+            }            
+            
+            $updated = "";
             if (isset($data['name'])) {
                 $author->setName($data['name']);
                 $updated .= " Name: " . $data['name'];
@@ -128,10 +154,9 @@ class Author extends BaseController {
                 $author->setIntroduction($data['introduction']);
                 $updated .= " Introduction: " . $data['introduction'];
             }
-            if (isset($data['user'])) {
-                $user = $this->getEntityManager()->find('Radio\Entity\User', $data['user']);
+            if ( !is_null($user) ) {
                 $author->setUser($user);
-                $updated .= " User: " . $data['user'];
+                $updated .= " User id: " . $data['user_id'];
             }
         
             $this->getEntityManager()->flush();
@@ -143,9 +168,12 @@ class Author extends BaseController {
     }
 
     public function delete($id) {
-        // TODO: validate via DB
         try {
             $author = $this->getEntityManager()->find('Radio\Entity\Author', $id);
+            if ( is_null($author) ) {
+                return new JsonModel(array("error" => "Author does not exist in DB."));
+            }
+            
             $this->getEntityManager()->remove($author);
             $this->getEntityManager()->flush();
